Fix build errors and warnings across platforms
Project-owned code:
- vm_x86.c: Add trailing comma to DROP() calls for -Wgnu-zero-variadic-macro-arguments
- sdl_icon.h: Add newline at end of file
- ui_css.c: Remove unused variable 'a'
- cl_cin_ffmpeg.c: Fix swr_convert pointer qualifier (pragma for safe cast)
- tr_main.c: Fix sign-compare (unsigned loop vars), cast-qual (const sortKey)
- tr_mesh.c: Use unsigned int for num_dlights loop
- tr_model.c: Cast sizeof for fileSize comparison
- tr_model_iqm.c: Fix IQM_CheckRange sign/const comparisons

External libs (CMake suppression):
- FLAC: -Wno-cast-qual -Wno-format-nonliteral -Wno-static-in-inline
- zlib: -Wno-format-nonliteral
- Duktape: -Wno-format-nonliteral -Wno-unused-but-set-variable
- Opus: -Wno-cast-align

+++ src/renderers/opengl/tr_model_iqm.c
@@ -39,9 +39,9 @@ static qboolean IQM_CheckRange( iqmHeader_t *header, int offset,
 	// doesn't fit into the file
 	return ( count <= 0 ||
 		 offset <= 0 ||
-		 offset > header->filesize ||
+		 (unsigned int)offset > header->filesize ||
 		 offset + count * size < 0 ||
-		 offset + count * size > header->filesize );
+		 (unsigned int)( offset + count * size ) > header->filesize );
 }
 // "multiply" 3x4 matrices, these are assumed to be the top 3 rows
 // of a 4x4 matrix with the last row = (0 0 0 1)
@@ -196,7 +196,7 @@ qboolean R_LoadIQM( model_t *mod, void *buffer, int filesize, const char *mod_na
 		float *f;
 	} blendWeights;
 
-	if( filesize < sizeof(iqmHeader_t) ) {
+	if( filesize < (int)sizeof(iqmHeader_t) ) {
 		return qfalse;
 	}
 
@@ -213,7 +213,7 @@ qboolean R_LoadIQM( model_t *mod, void *buffer, int filesize, const char *mod_na
 	}
 
 	LL( header->filesize );
-	if( header->filesize > filesize || header->filesize > 16<<20 ) {
+	if( header->filesize > (unsigned int)filesize || header->filesize > 16u<<20 ) {
 		return qfalse;
 	}
